requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
decorator==4.4.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
pbr==5.3.1
pycparser==2.19
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
